Move the location of the NCCO files, they need their own space
